move the game files to separate folder, and then exported to bit.dev
