+++ .bitmap
@@ -1,0 +1,38 @@
+/* THIS IS A BIT-AUTO-GENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. */
+
+{
+    "joshk.tile-15-game/game@0.0.1": {
+        "files": [
+            {
+                "relativePath": "src/components/game/game.component.css",
+                "test": false,
+                "name": "game.component.css"
+            },
+            {
+                "relativePath": "src/components/game/game.component.html",
+                "test": false,
+                "name": "game.component.html"
+            },
+            {
+                "relativePath": "src/components/game/game.component.ts",
+                "test": false,
+                "name": "game.component.ts"
+            },
+            {
+                "relativePath": "src/components/game/game.module.ts",
+                "test": false,
+                "name": "game.module.ts"
+            },
+            {
+                "relativePath": "src/components/game/index.ts",
+                "test": false,
+                "name": "index.ts"
+            }
+        ],
+        "mainFile": "src/components/game/index.ts",
+        "trackDir": "src/components/game",
+        "origin": "AUTHORED",
+        "exported": true
+    },
+    "version": "14.2.4"
+}
