+++ src/components/game/game.component.html
@@ -1,0 +1,20 @@
+<div class="puzzle-view">
+    <h1>The 15 Puzzle</h1>
+    <div class="puzzle">
+      <div class="tile-grid">
+        <div class="tile-row" *ngFor="let row of grid">
+          <div class="tile" *ngFor="let t of row">
+              <button disabled [ngClass]="{'empty': !success}" *ngIf="t === 0"><img style="width: 55px;" *ngIf="success" src="../assets/smiley.png"/></button>
+              <button id="{{t}}" [ngClass]="{'adjacent': !success}" *ngIf="t > 0 && adjacents.includes(t)" (click)="resetGrid(t)">{{t}}</button>
+              <button disabled *ngIf="t > 0 && !adjacents.includes(t)">{{t}}</button>
+          </div>
+        </div>
+      </div>
+      <h1 style="margin-top: 20px;color: rgb(93, 223, 158)" [hidden]="!success">Great Job! You solved the puzzle :)</h1>
+    </div>
+    <h1 class="instructions">How to Play:
+        <p>Select one of the yellow tiles to move it to the empty space</p>
+        <p>To win, order the tiles from smallest to largest starting at the top left</p>
+    </h1>
+  </div>
+
