+++ src/components/game/game.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { GameComponent } from './game.component';
+
+@NgModule({
+    imports: [CommonModule],
+    declarations: [GameComponent],
+    exports: [GameComponent]
+})
+export class GameModule { }
