+++ src/components/game/index.ts
@@ -1,0 +1,1 @@
+export { GameModule } from './game.module';
